Master

+++ inc/glpiuploadhandler.class.php
@@ -2,7 +2,7 @@
 /**
  * ---------------------------------------------------------------------
  * GLPI - Gestionnaire Libre de Parc Informatique
- * Copyright (C) 2015-2017 Teclib' and contributors.
+ * Copyright (C) 2015-2018 Teclib' and contributors.
  *
  * http://glpi-project.org
  *
@@ -30,10 +30,6 @@
  * ---------------------------------------------------------------------
  */
 
-/** @file
-* @brief
-*/
-
 if (!defined('GLPI_ROOT')) {
    die("Sorry. You can't access this file directly");
 }
@@ -42,7 +38,7 @@ include_once(GLPI_JQUERY_UPLOADHANDLER);
 
 /** GLPIUploadHandler class
  *
- * @since version 9.2
+ * @since 9.2
 **/
 class GLPIUploadHandler extends UploadHandler {
 
@@ -108,12 +104,26 @@ class GLPIUploadHandler extends UploadHandler {
             return __('Image requires a minimum height');
             break;
 
+         case 'accept_file_types':
+            return __('Filetype not allowed');
+            break;
+
+         case 'abort':
+            return __('File upload aborted');
+            break;
+
+         case 'image_resize':
+            return __('Failed to resize image');
+            break;
       }
 
       return false;
    }
 
    static function uploadFiles($params = []) {
+
+      global $DB;
+
       $default_params = [
          'name'           => '',
          'showfilesize'   => false,
@@ -121,32 +131,63 @@ class GLPIUploadHandler extends UploadHandler {
       ];
       $params = array_merge($default_params, $params);
 
+      $pname = $params['name'];
+      $rand_name = uniqid('', true);
+      foreach ($_FILES[$pname]['name'] as &$name) {
+         $name = $rand_name . $name;
+      }
+
+      $valid_type_iterator = $DB->request([
+         'FROM'   => 'glpi_documenttypes',
+         'WHERE'  => [
+            'is_uploadable'   => 1
+         ]
+      ]);
+
+      $valid_ext_patterns = [];
+      foreach ($valid_type_iterator as $valid_type) {
+         $valid_ext = $valid_type['ext'];
+         if (preg_match('/\/.+\//', $valid_ext)) {
+            // Filename matches pattern
+            // Remove surrounding '/' as it will be included in a larger pattern
+            // and protect by surrounding parenthesis to prevent conflict with other patterns
+            $valid_ext_patterns[] = '(' . substr($valid_ext, 1, -1) . ')';
+         } else {
+            // Filename ends with allowed ext
+            $valid_ext_patterns[] = '\.' . preg_quote($valid_type['ext'], '/') . '$';
+         }
+      }
+
       $upload_dir     = GLPI_TMP_DIR.'/';
-      $upload_handler = new self(['upload_dir'     => $upload_dir,
-                                  'param_name'     => $params['name'],
-                                  'orient_image'   => false,
-                                  'image_versions' => []],
-                                 false);
+      $upload_handler = new self(
+         [
+            'accept_file_types'         => '/(' . implode('|', $valid_ext_patterns) . ')/i',
+            'image_versions'            => [
+               'auto_orient' => false,
+            ],
+            'param_name'                => $pname,
+            'replace_dots_in_filenames' => false,
+            'upload_dir'                => $upload_dir,
+         ],
+         false
+      );
       $response       = $upload_handler->post(false);
 
       // clean compute display filesize
-      if (isset($response[$params['name']]) && is_array($response[$params['name']])) {
-         foreach ($response[$params['name']] as $key => &$val) {
-            if (Document::isValidDoc(addslashes($val->name))) {
+      if (isset($response[$pname]) && is_array($response[$pname])) {
+         foreach ($response[$pname] as &$val) {
+            if (isset($val->error) && file_exists($upload_dir.$val->name)) {
+               unlink($upload_dir.$val->name);
+            } else {
+               $val->prefix = $rand_name;
                if (isset($val->name)) {
-                  $val->display = $val->name;
+                  $val->display = str_replace($rand_name, '', $val->name);
                }
                if (isset($val->size)) {
                   $val->filesize = Toolbox::getSize($val->size);
                   if (isset($params['showfilesize']) && $params['showfilesize']) {
                      $val->display = sprintf('%1$s %2$s', $val->display, $val->filesize);
                   }
-               }
-            } else {
-               // Unlink file
-               $val->error = __('Filetype not allowed');
-               if (file_exists($upload_dir.$val->name)) {
-                  unlink($upload_dir.$val->name);
                }
             }
             $val->id = 'doc'.$params['name'].mt_rand();
@@ -155,5 +196,21 @@ class GLPIUploadHandler extends UploadHandler {
 
       // send answer
       return $upload_handler->generate_response($response, $params['print_response']);
+   }
+
+   protected function get_upload_data($id) {
+      return array_key_exists($id, $_FILES) ? $_FILES[$id] : null;
+   }
+
+   protected function get_post_param($id) {
+      return array_key_exists($id, $_POST) ? $_POST[$id] : null;
+   }
+
+   protected function get_query_param($id) {
+      return array_key_exists($id, $_GET) ? $_GET[$id] : null;
+   }
+
+   protected function get_server_var($id) {
+      return array_key_exists($id, $_SERVER) ? $_SERVER[$id] : null;
    }
 }
